+++ database/migrations/2016_11_16_120455_create_construction_daily_material_table.php
@@ -17,7 +17,7 @@ class CreateConstructionDailyMaterialTable extends Migration
             $table->increments('id');
             $table->integer('construction_daily_id')->unsigned();
             $table->integer('material_id')->unsigned();
-            $table->integer('amount')->unsigned();
+            $table->decimal('amount', 10, 2)->unsigned();
             $table->timestamps();
 
             $table->unique(['construction_daily_id', 'material_id'], 'ux_daily_material');
